refactor(db): Migrate clear database queries to SQLDelight

+++ app/src/main/java/eu/kanade/tachiyomi/data/database/models/SourceIdMangaCount.kt
@@ -1,3 +1,3 @@
 package eu.kanade.tachiyomi.data.database.models
 
-data class SourceIdMangaCount(val source: Long, val count: Int)
+data class SourceIdMangaCount(val source: Long, val count: Long)
